crypto.h : 1.	cambiado el nombre de la función para que sea con y griega 2.	iv y key cambiados de char a unsigned char
dataHolder
	cambiado size de long a int
crypto.c
	agregado. Por ahora solo funciona con DES

+++ trunk/wavEncrypt/src/dataHolder.c
@@ -4,7 +4,7 @@
 
 #include <stdlib.h>
 
-dataHolder_t newDataHolder(void *data, long size) {
+dataHolder_t newDataHolder(void *data, int size) {
 	dataHolder_t holder;
 
 	holder.data = data;
